Remove everything from Podman after each test
Otherwise, deleting the temporary test directory fails due to
permission issues. For example:
```
 Unable to delete /tmp/2316279192 on attempt #1, trying again...
  unlinkat /tmp/2316279192/containers/storage/overlay/d3ada5af5602bd0da378e4f0144f8fe8bdbe3f5a65f367dd9ebe759756bada68/diff/opt: permission deniedDeleting dir: /tmp/2316279192
```

Note that what is removed is only scoped to the current test storage location;
as such, it does not affect other tests or the global Podman storage
location on the system.
A future improvement would be to share the image store,
to speed up the tests a little bit by reusing images already pulled.
But the option to do so in storage.conf (`ImageStore`) has not been released yet.